% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_plot_mod.R
\name{display_plot_mod_server}
\alias{display_plot_mod_server}
\title{Plot variables Server module}
\usage{
display_plot_mod_server(id, var_inputs)
}
\arguments{
\item{id}{module id}

\item{var_inputs}{inputs from mod_var_input}
}
\value{
shiny server module
}
\description{
Plot variables Server module
}
