% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_plot_mod.R
\name{display_plot_mod_ui}
\alias{display_plot_mod_ui}
\title{Plot variables UI module}
\usage{
display_plot_mod_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny UI module
}
\description{
Plot variables UI module
}
