% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet_module.R
\name{greet_server_module}
\alias{greet_server_module}
\title{shinytest2 demo app}
\usage{
greet_server_module(id)
}
\arguments{
\item{input}{shiny parameter}

\item{output}{shiny parameter}

\item{session}{shiny parameter}
}
\value{
server for greeting app
}
\description{
This server function is for the simple greeting app
   used in the shinytest2 demonstration on YouTube:
   https://youtu.be/Gucwz865aqQ
}
