% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet_module.R
\name{greet_ui_module}
\alias{greet_ui_module}
\title{shinytest2 demo app}
\usage{
greet_ui_module(id)
}
\value{
UI for greeting app
}
\description{
This UI function is for the simple greeting app used in
   the shinytest2 demonstration on YouTube:
   https://youtu.be/Gucwz865aqQ
}
