% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_server}
\alias{mod_plot_server}
\title{Plot variables Server module}
\usage{
mod_plot_server(id, var_inputs)
}
\arguments{
\item{id}{}
}
\value{
shiny server module
}
\description{
Plot variables Server module
}
